split cache.py into classes
