+++ mb_modelbase/__init__.py
@@ -2,3 +2,4 @@ from mb_modelbase.models_core import *
 from mb_modelbase.model_eval import *
 from mb_modelbase.server import *
 from mb_modelbase.utils import *
+from mb_modelbase.cache import *
