+++ mb_modelbase/cache/__init__.py
@@ -1,0 +1,3 @@
+from .BaseCache import BaseCache
+from .DictCache import DictCache
+from .MemcachedCache import MemcachedCache
